[1343] Fixes the QA API
Fixes problems with the jsonification of dates, and calls that do not pass an id (but where it is required).

The id problem occurs where a route exists but the corresponding action function does not allow the id to be optional, and so in these cases the id in the params is defaulted to None and it 404s.

Where the json.dump calls fail because of dates I have used the ckan DateTime Json encoder to do the right thing.