[1367] Capture monthly broken pkg/rsc totals
Each month will capture (on the first of the month) the total number of broken datasets and resources for each publisher.  The data is stored in the DataCache and so is easily retrievable by the new API call.

As this part of the report is only generated on the 1st of the month, it might be necessary to regenerate this data on a day close to the first (for instance if the reports don't run for some reason).  To generate the report and always include the broken totals you should run with:

OVERRIDE_DATE_CHECK=1 paster cache-reports generate....

The new API call is available at /api/2/util/qa/broken_link_totals/<PUBLISHER_NAME> and returns a dict where the keys are sorted in date order. The dict for each key contains the total number of broken datasets/resources for that date.