AV-284 Single showcase page styling
- Showcase page had some issues with the pictures in the sidebar. These
were logos of other applications by the same user. These logos did not
fit well to our gray sidebar as the logos have no predefined size or
format specified (they can differ from the sites styling by quite a
lot). I decided to remove the imgaes from the sidebar altogether.
- The page had also some other issues with the applications logo in the
header that caused the title and other text to flow in very different
ways based on the logo size. Now the logo is on its own row and the
title and subheadings follow the logo after that.
- Other sidebar styling fixes.