AV-1461: Single apiset fails with internal server error
The data provided to the template is trying to access an index
from a list which might or might not be empty. Add a sanity check
to ensure that we only display the additional data section in the
apiset view if and only if the section contains any information.

Refs. AV-1461