AV-1509: Add translations to translations.py
Add missing translations to translations.py so the terms will be
picked up by the translator.

Add sentences for the "Show more" and "Show less" filter options
to the translations as well. These translations will use the
name of the facet rather than the title, in order to build
unique translatable strings in the template code for all of the
languages.

Refs AV-1509