AV-1474: Add active class on nav subpages
The navigation element should have the "is_active" css class on the element
if a user navigates to a page and its subpages.

Accomplish this by getting the base path used for the view, build a
full path by getting the alias of the base path and the language code.
Check if the url in the nav element exists in the full path, this will
essentially include subopages if they are under the same baseurl and the
is_active class will be set on the corresponding "parent" element.

Implement a helper function for getting the language code from the
$hook variable.

Refs. AV-1474