AV-1447: Remove users from organization
Implement script which allows a systems administrator to remove
users from an organization.

The users which are removed are either of member, admin or editor
capacity.

The script uses the -r flag for specifying the organization's
name or id from the command line. There is also a dry-run functionality
implemented which does not make any changes to the database but prints
the same information to stdout as it would when run normally.

Refs AV-1447